+++ src/main/kotlin/de/passbutler/desktop/ui/Intents.kt
@@ -5,3 +5,7 @@ import tornadofx.Component
 fun Component.openBrowser(url: String) {
     hostServices.showDocument(url)
 }
+
+fun Component.openWriteEmail(email: String) {
+    hostServices.showDocument("mailto:$email")
+}