+++ src/main/kotlin/de/passbutler/desktop/ui/NavigationMenuComponent.kt
@@ -1,6 +1,6 @@
 package de.passbutler.desktop.ui
 
-import de.passbutler.desktop.AboutScreen
+import de.passbutler.desktop.AboutLoggedInScreen
 import de.passbutler.desktop.ImportScreen
 import de.passbutler.desktop.OverviewScreen
 import de.passbutler.desktop.RecycleBinScreen
@@ -125,6 +125,6 @@ fun createDefaultNavigationMenu(): List<NavigationMenuUsing.NavigationItem> {
         NavigationMenuUsing.NavigationItem("drawer_menu_item_settings", Drawables.ICON_SETTINGS, SettingsScreen::class),
         NavigationMenuUsing.NavigationItem("drawer_menu_recycle_bin", Drawables.ICON_DELETE, RecycleBinScreen::class),
         NavigationMenuUsing.NavigationItem("drawer_menu_import", Drawables.ICON_MOVE_TO_INBOX, ImportScreen::class),
-        NavigationMenuUsing.NavigationItem("drawer_menu_item_about", Drawables.ICON_INFO, AboutScreen::class),
+        NavigationMenuUsing.NavigationItem("drawer_menu_item_about", Drawables.ICON_INFO, AboutLoggedInScreen::class),
     )
 }